refactor metric select and label filters, add variables to label names dropdown
